refactor: 디렉토리 경로 수정
- deal은 product라는 도메인의 하위 집합
