+++ src/app/_shared/components/MainView.tsx
