+++ src/app/product/(deal)/components/BrandProductList.tsx
