+++ src/app/product/(deal)/components/DealNavbar.tsx
