+++ src/app/product/(deal)/hooks/useDealTimer.tsx
