+++ src/app/product/[id]/components/AddToCartForm.tsx
@@ -15,10 +15,10 @@ import {
 import { Button } from "@/ui/button";
 import { Form } from "@/ui/form";
 
-import useAddToCartForm from "@/app/product/hooks/forms/useAddToCartForm";
+import useAddToCartForm from "@/app/product/[id]/hooks/forms/useAddToCartForm";
 
-import ProductOptions from "@/app/product/components/ProductOptions";
-import ProductQuantity from "@/app/product/components/ProductQuantity";
+import ProductOptions from "@/app/product/[id]/components/ProductOptions";
+import ProductQuantity from "@/app/product/[id]/components/ProductQuantity";
 
 type AddToCartFormProps = {
   productDetail: ProductDetailInfo;
